+++ src/Uno.Extensions.Maui.UI/GlobalUsings.cs
@@ -15,6 +15,7 @@ global using Microsoft.Maui.Embedding;
 global using Microsoft.Maui.Hosting;
 global using Microsoft.Maui.Platform;
 #if MAUI_EMBEDDING
+global using Uno.Extensions;
 global using Uno.Extensions.Maui.Internals;
 #endif
