+++ src/Uno.Extensions.Maui.UI/UnoServiceProviderFactory.cs
@@ -17,9 +17,30 @@ internal class UnoServiceProviderFactory : IServiceProviderFactory<IServiceProvi
 	public IServiceProvider CreateBuilder(IServiceCollection services)
 	{
 		// NOTE: TryAdd prevents duplicates from overriding what was registered with Uno's Host Builder such as IConfiguration
-		_mauiAppBuilder.Services.ForEach(x => services.TryAdd(x));
+		var instanceDescriptors = new List<ServiceDescriptor>();
+		_mauiAppBuilder.Services.ForEach(x =>
+		{
+			if (!services.TryAddService(x) )
+			{
+				if (x.ImplementationInstance is not null)
+				{
+					instanceDescriptors.Add(x);
+				}
+			}
+		});
 
 		var serviceProvider = services.BuildServiceProvider();
+		instanceDescriptors.ForEach(x =>
+		{
+			if (serviceProvider is not null &&
+				x.ImplementationInstance is not null)
+			{
+				var implType = x.ImplementationType ?? x.ImplementationInstance.GetType();
+
+				serviceProvider.AddSingletonInstance(implType, x.ImplementationInstance);
+			}
+		});
+
 		_mauiAppBuilder.ConfigureContainer(new MauiServiceProviderFactory(serviceProvider));
 		_buildAppBuilderCallback();
 		return serviceProvider;
